requirements.txt
aiohttp==1.0.5
async-timeout==3.0.0
certifi==2018.8.13
chardet==3.0.4
discord.py==0.16.12
multidict==4.3.1
pipenv==2018.7.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
websockets==3.4
